+++ README.md
@@ -1,21 +1,30 @@
-[B]4.0.5[/B]
+Missing Movies KODI add-on
+=============================
+Scans media sources for movies, TV shows and episodes that got missed during the library scan. This add-on can used to identify, play or manually add missing videos to your media library.                            
+
+Changelog
+
+**4.1.0**
+- Ported to python 3.0 to run on v19 Matrix
+
+**4.0.5**
 - Improved detection of bluray companion files.
 
-[B]4.0.4[/B]
+**4.0.4**
 - Fixed detection of stacked files with comma in filename.
 
-[B]4.0.3[/B]
+**4.0.3**
 - Fixed case conversion for blacklists (now case insensitive)
 
-[B]4.0.2[/B]
+**4.0.2**
 - Fixed multi-path sources
 - Fixed bluray backup files detected as missing
 - Added .url files to the default blacklist
 
-[B]4.0.1[/B]
+**4.0.1**
 - Fixed remote paths getting identified as missing on windows
 
-[B]4.0.0[/B]
+**4.0.0**
 - Reworked existing features
 - Added ability to show identify missing TV shows
 - Improved detection of DVD/Bluray files, samples, extras and other false positives
