+++ addon.xml
@@ -1,18 +1,29 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="plugin.video.mms" name="Missing Movies" version="4.0.5" provider-name="takoi">
+<addon id="plugin.video.mms" name="Missing Movies" version="4.1.0" provider-name="takoi, juokelis">
     <requires>
-        <import addon="xbmc.python" version="2.14.0"/>
-        <import addon="script.module.routing" version="0.2.0"/>
+        <import addon="xbmc.python" version="3.0.0"/>
+        <import addon="script.module.routing" version="0.2.3+matrix.1"/>
     </requires>
     <extension point="xbmc.python.pluginsource" library="default.py">
         <provides>video</provides>
     </extension>
     <extension point="xbmc.addon.metadata">
-        <summary lang="en">Find missing videos</summary>
-        <description lang="en">Scans media sources for movies, TV shows and episodes that got missed during the library scan. This add-on can used to identify, play or manually add missing videos to your media library.</description>
-        <license>GNU GENERAL PUBLIC LICENSE Version 3, 29 June 2007</license>
+        <summary lang="en_GB">Find missing videos</summary>
+        <description lang="en_GB">Scans media sources for movies, TV shows and episodes that got missed during the library scan. This add-on can used to identify, play or manually add missing videos to your media library.</description>
+        <news>
+v4.1.0 (2020-04-05) by juokelis
+- Ported to python 3.0 to run on v19 Matrix.
+
+v4.0.5 (2016-01-17) by tamland
+- Improved detection of bluray companion files.
+        </news>
+        <license>GPL-3.0-or-later</license>
         <platform>all</platform>
         <source>https://github.com/tamland/kodi-missing-movies</source>
-        <forum>http://forum.xbmc.org/showthread.php?tid=118263</forum>
+        <forum>https://forum.kodi.tv/showthread.php?tid=118263</forum>
+        <assets>
+            <icon>resources/media/icon.png</icon>
+            <fanart>resources/media/fanart.jpg</fanart>
+        </assets>
     </extension>
 </addon>
